Refactor notification banner
