+++ src/components/info/banners/notification_banner.tsx
