+++ src/components/routes/Home.tsx
@@ -10,8 +10,8 @@ const FooterComponent = lazy(() => import(/* webpackChunkName: 'footer' */ 'comp
 const LoadedArmyHeader = lazy(() =>
   import(/* webpackChunkName: 'loaded_army_header' */ 'components/input/savedArmies/loaded_army_header')
 )
-const OrrukCoSBanner = lazy(() =>
-  import(/* webpackChunkName: 'orruk_cos_banner' */ 'components/info/banners/orruk_cos_banner')
+const AppBanner = lazy(() =>
+  import(/* webpackChunkName: 'app_banner' */ 'components/info/banners/app_banner')
 )
 const PrintArmy = lazy(() => import(/* webpackChunkName: 'printArmy' */ 'components/print/printArmy'))
 const PrintFooter = lazy(() => import(/* webpackChunkName: 'printFooter' */ 'components/print/printFooter'))
@@ -34,7 +34,7 @@ const Home: React.FC = () => {
       <Header />
 
       <Suspense fallback={<></>}>
-        <OrrukCoSBanner />
+        <AppBanner />
       </Suspense>
 
       <Suspense fallback={<></>}>
